Address Seb's feedback and fix tests. - Combine source field and dynamic mapping forms into a single configuration form. - Use nested schema. - Refactor default value construction and configuration form de/serializer functions to be explicit about how inputs are transformed to outputs.
