+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/configuration_form.tsx
@@ -3,30 +3,90 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import React from 'react';
-
+import React, { useEffect } from 'react';
 import { EuiSpacer } from '@elastic/eui';
 
-import { Types } from '../../mappings_state';
-import { DynamicMappingForm } from './dynamic_mapping_form';
-import { SourceFieldForm } from './source_field_form';
+import { useForm, Form, SerializerFunc } from '../../shared_imports';
+import { Types, useDispatch } from '../../mappings_state';
+import { DynamicMappingSection } from './dynamic_mapping_section';
+import { SourceFieldSection } from './source_field_section';
+import { configurationFormSchema } from './configuration_form_schema';
 
 type MappingsConfiguration = Types['MappingsConfiguration'];
-type SourceField = Types['SourceField'];
 
 interface Props {
-  configurationDefaultValue?: MappingsConfiguration;
-  sourceFieldDefaultValue?: SourceField;
+  defaultValue?: MappingsConfiguration;
 }
 
-export const ConfigurationForm = React.memo(
-  ({ configurationDefaultValue, sourceFieldDefaultValue }: Props) => {
-    return (
-      <>
-        <DynamicMappingForm defaultValue={configurationDefaultValue} />
-        <EuiSpacer size="xl" />
-        <SourceFieldForm defaultValue={sourceFieldDefaultValue} />
-      </>
-    );
-  }
-);
+const formSerializer: SerializerFunc<MappingsConfiguration> = formData => {
+  const {
+    dynamicMapping: {
+      enabled: dynamicMappingsEnabled,
+      throwErrorsForUnmappedFields,
+      numeric_detection,
+      date_detection,
+      dynamic_date_formats,
+    },
+    sourceField,
+  } = formData;
+
+  const dynamic = dynamicMappingsEnabled ? true : throwErrorsForUnmappedFields ? 'strict' : false;
+
+  return {
+    dynamic,
+    numeric_detection,
+    date_detection,
+    dynamic_date_formats,
+    _source: { ...sourceField },
+  };
+};
+
+const formDeserializer = (formData: { [key: string]: any }) => {
+  const {
+    dynamic,
+    numeric_detection,
+    date_detection,
+    dynamic_date_formats,
+    _source: { enabled, includes, excludes },
+  } = formData;
+
+  return {
+    dynamicMapping: {
+      enabled: dynamic === true || dynamic === undefined,
+      throwErrorsForUnmappedFields: dynamic === 'strict',
+      numeric_detection,
+      date_detection,
+      dynamic_date_formats,
+    },
+    sourceField: {
+      enabled: enabled === true || enabled === undefined,
+      includes,
+      excludes,
+    },
+  };
+};
+
+export const ConfigurationForm = React.memo(({ defaultValue }: Props) => {
+  const { form } = useForm<MappingsConfiguration>({
+    schema: configurationFormSchema,
+    serializer: formSerializer,
+    deserializer: formDeserializer,
+    defaultValue,
+  });
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const subscription = form.subscribe(updatedConfiguration => {
+      dispatch({ type: 'configuration.update', value: updatedConfiguration });
+    });
+    return subscription.unsubscribe;
+  }, [form]);
+
+  return (
+    <Form form={form}>
+      <DynamicMappingSection />
+      <EuiSpacer size="xl" />
+      <SourceFieldSection />
+    </Form>
+  );
+});
