--- x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/dynamic_mapping_form/dynamic_mapping_schema.tsx
@@ -1,104 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-import React from 'react';
-
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
-import { EuiLink } from '@elastic/eui';
-
-import { documentationService } from '../../../../../services/documentation';
-import {
-  FormSchema,
-  FIELD_TYPES,
-  VALIDATION_TYPES,
-  fieldValidators,
-} from '../../../shared_imports';
-import { MappingsConfiguration } from '../../../reducer';
-
-const { containsCharsField } = fieldValidators;
-
-export const dynamicMappingSchema: FormSchema<MappingsConfiguration> = {
-  enabled: {
-    label: i18n.translate('xpack.idxMgmt.mappingsEditor.configuration.enableDynamicMappingsLabel', {
-      defaultMessage: 'Enable dynamic mapping',
-    }),
-    type: FIELD_TYPES.TOGGLE,
-    defaultValue: true,
-  },
-  throwErrorsForUnmappedFields: {
-    label: i18n.translate(
-      'xpack.idxMgmt.mappingsEditor.configuration.throwErrorsForUnmappedFieldsLabel',
-      {
-        defaultMessage: 'Throw an exception when a document contains an unmapped field',
-      }
-    ),
-    helpText: i18n.translate('xpack.idxMgmt.mappingsEditor.dynamicMappingStrictHelpText', {
-      defaultMessage:
-        'By default, unmapped fields will be silently ignored when dynamic mapping is disabled. Optionally, you can choose to throw an exception when a document contains an unmapped field.',
-    }),
-    type: FIELD_TYPES.CHECKBOX,
-    defaultValue: false,
-  },
-  numeric_detection: {
-    label: i18n.translate('xpack.idxMgmt.mappingsEditor.configuration.numericFieldLabel', {
-      defaultMessage: 'Map numeric strings as numbers',
-    }),
-    helpText: i18n.translate('xpack.idxMgmt.mappingsEditor.configuration.numericFieldDescription', {
-      defaultMessage:
-        'For example, "1.0" would be mapped as a float and "1" would be mapped as an integer.',
-    }),
-    type: FIELD_TYPES.TOGGLE,
-    defaultValue: false,
-  },
-  date_detection: {
-    label: i18n.translate('xpack.idxMgmt.mappingsEditor.configuration.dateDetectionFieldLabel', {
-      defaultMessage: 'Map date strings as dates',
-    }),
-    type: FIELD_TYPES.TOGGLE,
-    defaultValue: true,
-  },
-  dynamic_date_formats: {
-    label: i18n.translate('xpack.idxMgmt.mappingsEditor.configuration.dynamicDatesFieldLabel', {
-      defaultMessage: 'Date formats',
-    }),
-    helpText: () => (
-      <FormattedMessage
-        id="xpack.idxMgmt.mappingsEditor.configuration.dynamicDatesFieldHelpText"
-        defaultMessage="Strings in these formats will be mapped as dates. You can use built-in formats or custom formats. {docsLink}"
-        values={{
-          docsLink: (
-            <EuiLink href={documentationService.getDateFormatLink()} target="_blank">
-              {i18n.translate(
-                'xpack.idxMgmt.mappingsEditor.configuration.dynamicDatesFieldDocumentionLink',
-                {
-                  defaultMessage: 'Learn more.',
-                }
-              )}
-            </EuiLink>
-          ),
-        }}
-      />
-    ),
-    type: FIELD_TYPES.COMBO_BOX,
-    defaultValue: ['strict_date_optional_time', 'yyyy/MM/dd HH:mm:ss Z||yyyy/MM/dd Z'],
-    validations: [
-      {
-        validator: containsCharsField({
-          message: i18n.translate(
-            'xpack.idxMgmt.mappingsEditor.configuration.dynamicDatesFieldValidationErrorMessage',
-            {
-              defaultMessage: 'Spaces are not allowed.',
-            }
-          ),
-          chars: ' ',
-        }),
-        type: VALIDATION_TYPES.ARRAY_ITEM,
-      },
-    ],
-  },
-};
-
-export const CONFIGURATION_FIELDS = Object.keys(dynamicMappingSchema);
