+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/index.ts
@@ -5,4 +5,3 @@
  */
 
 export { ConfigurationForm } from './configuration_form';
-export { CONFIGURATION_FIELDS } from './dynamic_mapping_form';
